Add packaging Dockerfiles for various distribution and Python version combos (#477)
* Set up symlink from centos8 -> centos.

* Add Dockerfiles for centos6 and 7 with Python 2 and Python 3 install.

This way we can use those Dockerfiles to test various different Python
version scenarios.

* Add ubuntu1604 Docker file we can use for testing things.

* Add various distro combinations.

* Fix agent runner so it works under Python 2 only environment.

* Update existing files and add support for running those tests under
Ubuntu 16.04 and CentOS 7.

* Simplify postinstall scripts

* Fix ordering of redirect

* Run different set of tests under CentOS 7 where python binary is
available.

* Don't throw inside _run_command but inside the caller function.

* Handle a scenario where service executable is not available and use full
absolute path when it is.

* Add a workaround for distros where a reverse symlink is set up (e.g.
python2 -> python instead of python -> python2).

* Add a workaround for service command on CentOS.

We need to use absolute path there and can't use
/etc/init.d/scalyr-agent-2 because it will hang.

* Don't run problematic tests under centos6 so we can simplify the code.

* Update dockerized decorator so it correctly handles artifacts copying.

We run each test case in a separate container which means we need to
preserve log file for each test case run (aka each container).

Previous code didn't take this into account so it only copied /
preserved agent.log file from the container which belonged to the last
test run.

In addition to that, update the code to support preservation of
arbitrary number of file paths.

* Update test files so we preserve built rpm and deb package artifacts.

This allows developers to easily use and tests artifacts from those builds.
* Update install deb function so it also works with older versions on apt
(aka on Ubuntu 14.04).

* Add file for running tests on Ubuntu 14.04.


* Reduce Dockerfile content duplication and move more of the duplicated
content in a base deb and rpm Dockerfile.

Also update agent version to include git revision.

* Add a fix / workaround for installation failure on Ubuntu 18.04 which
only ships with trimmed down Python 3 version which lacks
distutils.spawn package.

* Add a bug fix for rc*.d symlink creation failure on Ubuntu versions with
systemd.

* Add unit test which ensures agent_main.py contains correct init pragma
comment format.

* Add new simpler tox and Circle CI jobs which only build agent deb and
rpm packages.

This way we can speed things up by avoiding running tests.

* Make sure we also record Circle CI usage data for new workflow and
update the script to include stats for master branch first (since that's
the one we primarily care about).

+++ tests/distribution/python_version_change_tests/common.py
@@ -32,8 +32,11 @@ from tests.utils.compat import Path
 from tests.utils.common import get_shebang_from_file
 from tests.utils.agent_runner import AgentRunner, PACKAGE_INSTALL
 from tests.common import PackageInstallationError
+from tests.common import install_deb, remove_deb
 
 SCALYR_PACKAGE_BIN_PATH = Path("/", "usr", "share", "scalyr-agent-2", "bin")
+
+# NOTE: Binary dir path is different across distros that's why we use which to locate it
 BINARY_DIR_PATH = Path("/", "usr", "bin")
 
 
@@ -67,12 +70,16 @@ def _link_to_default_python(command):
     :return:
     """
     python_path = six.text_type(BINARY_DIR_PATH / "python")
+    real_executable_path = os.readlink(six.text_type(BINARY_DIR_PATH / command))
+
+    if _python_binary_is_symlink() and command == "python2":
+        return
+
     try:
         os.remove(python_path)
     except:
         pass
 
-    real_executable_path = os.readlink(six.text_type(BINARY_DIR_PATH / command))
     os.symlink(six.text_type(BINARY_DIR_PATH / real_executable_path), python_path)
 
 
@@ -98,7 +105,7 @@ def _mock_python_binary_version(python_binary_name, version):
     if not binary_path.exists():
         return
 
-    # make backup of the original binary in case if we want to keep useing it.
+    # make backup of the original binary in case if we want to keep using it.
     shutil.copy(six.text_type(binary_path), six.text_type(binary_path_backup_path))
     os.remove(six.text_type(binary_path))
 
@@ -114,6 +121,36 @@ def _mock_binaries(python, python2, python3):
     _mock_python_binary_version("python", python)
     _mock_python_binary_version("python2", python2)
     _mock_python_binary_version("python3", python3)
+
+
+def common_test_ubuntu_versions():
+    runner = AgentRunner(PACKAGE_INSTALL)
+    common_version_test(
+        runner,
+        install_deb,
+        remove_deb,
+        None,
+        "2.5.1",
+        "2.5.1",
+        "3.4.1",
+        install_fails=True,
+    )
+    common_version_test(
+        runner, install_deb, remove_deb, "config_main.py", "", "2.5.1", "3.4.1"
+    )
+    common_version_test(
+        runner, install_deb, remove_deb, "config_main_py2.py", "2.5.1", "", "3.4.1"
+    )
+    common_version_test(
+        runner, install_deb, remove_deb, "config_main_py3.py", "2.5.1", "2.5.1", ""
+    )
+    common_version_test(runner, install_deb, remove_deb, "config_main.py", "", "", "")
+    common_version_test(
+        runner, install_deb, remove_deb, "config_main_py2.py", "2.5.1", "", ""
+    )
+    common_version_test(
+        runner, install_deb, remove_deb, "config_main.py", "", "2.5.1", ""
+    )
 
 
 def common_version_test(
@@ -147,7 +184,9 @@ def common_version_test(
     else:
         stdout, _ = install_package_fn()
 
-    assert _get_current_config_script_name() == expected_conf_file_name
+    current_config_script_file_name = _get_current_config_script_name()
+
+    assert current_config_script_file_name == expected_conf_file_name
 
     _mock_binaries("", "", "")
 
@@ -183,10 +222,12 @@ def common_test_only_python_mapped_to_python2(
     Test package installation on the machine with python2 but there is only 'python' command which is mapped on to it.
     :param install_package_fn: callable that installs package with appropriate type to the current machine OS.
     """
-
     # map 'python' command on to python2
     _link_to_default_python("python2")
-    _remove_python("python2")
+    if not _python_binary_is_symlink():
+        # On some older distros python2 is mapped to python and not vice-versa so we only remove
+        # that binary if that is not the case
+        _remove_python("python2")
     _remove_python("python3")
 
     stdout, _ = install_package_fn()
@@ -254,7 +295,6 @@ def common_test_python2(install_package_fn, install_next_version_fn):
     :param install_package_fn: callable that installs package with appropriate type to the current machine OS.
     """
 
-    # remove python and python3 to make installer see only python2
     _remove_python("python")
     _remove_python("python3")
 
@@ -518,7 +558,6 @@ def common_test_switch_python2_to_python3(install_package_fn, install_next_versi
     Package installer should pick python2 by default and then we switch to the python3.
     :param install_package_fn: callable that installs package with appropriate type to the current machine OS.
     """
-
     _remove_python("python")
 
     install_package_fn()
@@ -552,3 +591,22 @@ def common_test_switch_python2_to_python3(install_package_fn, install_next_versi
     runner.start()
     time.sleep(1)
     assert _get_python_major_version(runner) == 2
+
+
+def _python_binary_is_symlink():
+    # type: () -> bool
+    """
+    Return true if python binary is a symlink to a specific version (e.g. python -> python2 instead
+    of the more usal python2 -> python).
+
+    This affects our tests because we need to handle those scenarios separately.
+    """
+    try:
+        os.readlink(six.text_type(BINARY_DIR_PATH / "python"))
+    except OSError as e:
+        msg = str(e).lower()
+
+        if "invalid argument" in msg:
+            return True
+
+    return False
